feat(test) : [추가] BuyerTest
